move stuff in compression around
