+++ compression/src/main/java/net/daporkchop/lib/compression/util/StreamingWrapperCCtx.java
@@ -26,10 +26,9 @@ import lombok.Getter;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.Accessors;
-import net.daporkchop.lib.compression.CCtx;
-import net.daporkchop.lib.compression.CompressionProvider;
-import net.daporkchop.lib.compression.PDeflater;
-import net.daporkchop.lib.compression.StreamingCompressionProvider;
+import net.daporkchop.lib.compression.context.CCtx;
+import net.daporkchop.lib.compression.context.PDeflater;
+import net.daporkchop.lib.compression.provider.StreamingCompressionProvider;
 import net.daporkchop.lib.compression.util.exception.DictionaryNotAllowedException;
 import net.daporkchop.lib.compression.util.exception.InvalidCompressionLevelException;
 import net.daporkchop.lib.natives.util.exception.InvalidBufferTypeException;
