+++ compression/zlib/src/main/java/net/daporkchop/lib/compression/zlib/ZlibCCtx.java
@@ -20,10 +20,7 @@
 
 package net.daporkchop.lib.compression.zlib;
 
-import io.netty.buffer.ByteBuf;
-import lombok.NonNull;
-import net.daporkchop.lib.compression.CCtx;
-import net.daporkchop.lib.natives.util.exception.InvalidBufferTypeException;
+import net.daporkchop.lib.compression.context.CCtx;
 
 /**
  * An extension of {@link CCtx} for {@link Zlib}.
