+++ compression/zlib/src/main/java/net/daporkchop/lib/compression/zlib/ZlibDeflater.java
@@ -22,7 +22,7 @@ package net.daporkchop.lib.compression.zlib;
 
 import io.netty.buffer.ByteBuf;
 import lombok.NonNull;
-import net.daporkchop.lib.compression.PDeflater;
+import net.daporkchop.lib.compression.context.PDeflater;
 import net.daporkchop.lib.natives.util.exception.InvalidBufferTypeException;
 
 /**
