+++ compression/zlib/src/main/java/net/daporkchop/lib/compression/zlib/ZlibInflater.java
@@ -22,7 +22,7 @@ package net.daporkchop.lib.compression.zlib;
 
 import io.netty.buffer.ByteBuf;
 import lombok.NonNull;
-import net.daporkchop.lib.compression.PInflater;
+import net.daporkchop.lib.compression.context.PInflater;
 import net.daporkchop.lib.natives.util.exception.InvalidBufferTypeException;
 
 /**
