+++ compression/zlib/src/main/java/net/daporkchop/lib/compression/zlib/natives/NativeZlibCCtx.java
@@ -20,14 +20,11 @@
 
 package net.daporkchop.lib.compression.zlib.natives;
 
-import io.netty.buffer.ByteBuf;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.experimental.Accessors;
-import net.daporkchop.lib.compression.CCtx;
 import net.daporkchop.lib.compression.util.StreamingWrapperCCtx;
 import net.daporkchop.lib.compression.zlib.ZlibCCtx;
-import net.daporkchop.lib.natives.util.exception.InvalidBufferTypeException;
 
 /**
  * @author DaPorkchop_
