+++ compression/zlib/src/main/java/net/daporkchop/lib/compression/zlib/natives/NativeZlibInflater.java
@@ -25,7 +25,7 @@ import lombok.Getter;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.Accessors;
-import net.daporkchop.lib.compression.PInflater;
+import net.daporkchop.lib.compression.context.PInflater;
 import net.daporkchop.lib.compression.util.exception.ContextFinishedException;
 import net.daporkchop.lib.compression.util.exception.ContextFinishingException;
 import net.daporkchop.lib.compression.zlib.ZlibInflater;
