+++ compression/zstd/src/main/java/net/daporkchop/lib/compression/zstd/ZstdCCtx.java
@@ -22,7 +22,7 @@ package net.daporkchop.lib.compression.zstd;
 
 import io.netty.buffer.ByteBuf;
 import lombok.NonNull;
-import net.daporkchop.lib.compression.CCtx;
+import net.daporkchop.lib.compression.context.CCtx;
 import net.daporkchop.lib.compression.util.exception.DictionaryNotAllowedException;
 import net.daporkchop.lib.natives.util.exception.InvalidBufferTypeException;
